% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somma.R
\name{somma}
\alias{somma}
\title{somma}
\usage{
somma(a, b)
}
\arguments{
\item{a}{\link{numeric} first addendum}

\item{b}{\link{numeric} second addendum}
}
\value{
\link{numeric} the sum of the two addendum
}
\description{
This function sums two numbers a^2
}
\examples{
somma(2,3)
}
